All simple types are now passed around inside of DSHandle objects, instead of sometimes inside DSHandles and sometimes as rawer Java objects.
Swift operators now implemented in their own class, rather than using
the Karajan implementation.

There are now two division operators, / and %/ of which the first (/)
returns a floating point value and %/ returns an integer value.
The mod operator is relabelled %% instead of %.

(this is mostly a merge of work from benc and hategan on a branch)



git-svn-id: https://svn.ci.uchicago.edu/svn/vdl2/trunk@1008 e2bb083e-7f23-0410-b3a8-8253ac9ef6d8