the parser doesn't need to know about simple types explicitly - type handling there is subsumed into the arbitrary type name handling.
as part of this change, the use of xs: prefix on types is removed.
that changes what the XML schema declarations looks like in .xml files - 
strictly interpreting that XML scheme, we are now no longer declaring
the XML base types called int or float or string; instead we are declaring
types in the swift namespace with the same local parts.
However, there's no namespace handling for types so that doesn't matter
in implementation.


git-svn-id: https://svn.ci.uchicago.edu/svn/vdl2/trunk@1537 e2bb083e-7f23-0410-b3a8-8253ac9ef6d8