Pass source file line numbers deeper into the compiler, and make use of them to give location information in more error messages.
Every time an XDTM.stg template is created, it will be given the present
source location as a sourcelocation parameter. This is then used in some
(but by no means all) of the XML intermediate forms to populate a src
attribute. At XML->Karajan compile time, this is then used to give the
original source location in compilation error messages.



git-svn-id: https://svn.ci.uchicago.edu/svn/vdl2/trunk@1691 e2bb083e-7f23-0410-b3a8-8253ac9ef6d8