Change how trace() deals with non-primitive datasets:
* Array handling:
Previous to this, trace would show the Java object representation of a (Future)PairIterator for arrays, which is fairly useless for a user. This commit makes trace show more about the array, and only emit the trace when the array is closed.

* Other datasets:
Trace will wait for those datasets to be closed, and emit their default
string representation (including the variable name and path used in
SwiftScript)

git-svn-id: https://svn.ci.uchicago.edu/svn/vdl2/trunk@2735 e2bb083e-7f23-0410-b3a8-8253ac9ef6d8