Various fixes to the test suite script
The current suite.sh does not kill all processes correctly. When a swift script exceeds its time limit, the parent bash script gets a kill signal, but the children processes do not, which leaves many swift/java instances running after the suite is over. This should be fixed now.

Previously suite.sh would attempt to determine the IP address of the local machine via ifconfig and set GLOBUS_HOSTNAME. It's not easy to determine which IP address is the correct one to use. On macs it fails completely and sets it to "inet". It's probably better to let the user define it, either globally via .bashrc, or within the user environment script (-z user_env.sh).

Replace readlink command for determining directory structures - it failed on macs and caused the suite to not run. The test suite should be runnable on macs now.

The repeat file was not being read correctly. This should be working now for repeating tests.



git-svn-id: https://svn.ci.uchicago.edu/svn/vdl2/branches/release-0.94@7050 e2bb083e-7f23-0410-b3a8-8253ac9ef6d8